Cloud-LLM-Apps/EndtoendRAGusingAmazonBedrockFaissProject/requirements.txt
langchain
langchain-community
langchain-aws
langchain-text-splitters
langchain-core
pypdf
streamlit
faiss-cpu
boto3
python-dotenv
